requirements.txt
tslearn
holidays
matplotlib
pandas
openpyxl
numpy
numpy_financial
h5py
scikit-learn
seaborn
PyYAML
holidays
pyemd
